+++ src/main/java/com/nyokinyoki/Attend/AttendStatus.java
@@ -1,6 +1,6 @@
 package com.nyokinyoki.Attend;
 
-import java.time.*;
+import java.time.LocalDate;
 
 import com.nyokinyoki.Timetable.Timeslot;
