+++ src/main/java/com/nyokinyoki/IntegrationTest.java
@@ -1,16 +1,17 @@
 package com.nyokinyoki;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 import com.nyokinyoki.Attend.AttendManager;
 import com.nyokinyoki.Attend.AttendStatus;
-import com.nyokinyoki.Course.Course;
+import com.nyokinyoki.DAO.CourseDAO;
 import com.nyokinyoki.Timestamp.StampStatus;
-import com.nyokinyoki.Timetable.TimeCard;
-import com.nyokinyoki.Timetable.Timetable;
-
-import java.time.*;
-import java.time.format.*;;
+import com.nyokinyoki.Timestamp.TimeCard;
+import com.nyokinyoki.Timetable.Course;
+import com.nyokinyoki.Timetable.Timetable;;
 
 public class IntegrationTest {
