+++ src/main/java/com/nyokinyoki/Terminal/CLIArg.java
@@ -1,6 +1,7 @@
 package com.nyokinyoki.Terminal;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
 
 import org.jline.utils.AttributedString;
